% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dots.R
\name{calc_dots}
\alias{calc_dots}
\title{calc_dots}
\usage{
calc_dots(df, col_names, n_per_dot)
}
\arguments{
\item{df}{the merged df of a shapefile and population data}

\item{col_names}{a vector of col_names to select from this merged data. If selecting all columns, can leave as NULL}

\item{n_per_dot}{the number of n people in each category for every dot}
}
\description{
A function which calculates the position of number and position
    of dots in the final plot for each row of data.
}
\details{
credit to Paul Campbell for the function in his blogpost https://www.cultureofinsight.com/blog/2018/05/02/2018-04-08-multivariate-dot-density-maps-in-r-with-sf-ggplot2/
}
\author{
Paul Campbell, Robert Hickman
}
