% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_data.R
\docType{data}
\name{ge_data}
\alias{ge_data}
\title{Electoral Data from Paul Campbell's blogpost}
\format{A data frame with 73 rows and 8 variables}
\usage{
data(ge_data)
}
\description{
A dataset containing the results of the 2017 UK General Election
 with votes for major parties by the seat in the House of Commons
}
\details{
\itemize{
  \item ons_id. The Office of National Statistics id for the constituency
  \item constituency_name. The plain language name of the constituency seat
  \item first_party. The parliamentary party with the most votes for a given seat
  \item Con. The number of votes won by the Conservative party for a given seat
  \item Lab. The number of votes won by the Labour party for a given seat
  \item LD. The number of votes won by the Liberal Democrats party for a given seat
  \item UKIP. The number of votes won by the UKIP party for a given seat
  \item Green. The number of votes won by the Green party for a given seat
}
}
\keyword{datasets}
