% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_round.R
\name{random_round}
\alias{random_round}
\title{random_round}
\usage{
random_round(x)
}
\description{
A function which calculates how many dots each class
will have for each given shape
}
\details{
credit to Jens von Bergmann for this algo https://github.com/mountainMath/dotdensity/blob/master/R/dot-density.R
}
\author{
Jens von Bergmann, Paul Campbell, Robert Hickman
}
