% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\docType{data}
\name{uk}
\alias{uk}
\title{Electoral Data from Paul Campbell's blogpost}
\format{A data frame with 73 rows and 2 variables}
\usage{
data(uk)
}
\description{
A dataset containing the simple features shapefile for the London
    electoral seats in the 2017 UK General Election
}
\details{
\itemize{
  \item ons_id. The Office of National Statistics id for the constituency
  \item geometery. The geometry of the shapefile
}
}
\keyword{datasets}
