% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosnia_ethnicity_data.R,
%   R/london_election_data.R
\docType{data}
\name{london_election_data}
\alias{london_election_data}
\alias{bosnia_ethnicity_data}
\title{Census Ethnicity Data to replicate Bosnia dot density map from https://twitter.com/ImerM1/status/1037358973807210498}
\format{A data frame with 141 rows and 20 variables}
\usage{
data(london_election_data)

data(london_election_data)
}
\description{
A dataset containing the ethnicity data for Bosnian districts in
 the 2013 census

A dataset containing the results of the 2017 UK General Election in London
 with votes for major parties by the seat in the House of Commons
}
\details{
\itemize{
  \item{district}{The name of the district for the 2013 Bosnian census}
  \item{Total}{The number of people of all ethnicities in the 2013 Bosnian census}
  \item{Bosniak}{The number of people of Bosniak ethnicity in the 2013 Bosnian census}
  \item{Serb}{The number of people of Serb ethnicity in the 2013 Bosnian census}
  \item{Croat}{The number of people of Croat ethnicity in the 2013 Bosnian census}
  \item{Bosnian}{The number of people of Bosnian ethnicity in the 2013 Bosnian census}
  \item{Romani}{The number of people of Romani ethnicity in the 2013 Bosnian census}
  \item{Muslim}{The number of people of Muslim ethnicity in the 2013 Bosnian census}
  \item{Bosnian.and.Herzegovinian}{The number of people of Bosnian and Herzegovinian ethnicity in the 2013 Bosnian census}
  \item{Albanian}{The number of people of Albanian ethnicity in the 2013 Bosnian census}
  \item{Yugoslav}{The number of people of Yugoslav ethnicity in the 2013 Bosnian census}
  \item{Ukranian}{The number of people of Ukranian ethnicity in the 2013 Bosnian census}
  \item{Montenegrin}{The number of people of Montenegrin ethnicity in the 2013 Bosnian census}
  \item{Turk}{The number of people of Turkish ethnicity in the 2013 Bosnian census}
  \item{Slovenian}{The number of people of Slovenian ethnicity in the 2013 Bosnian census}
  \item{Orthodox}{The number of people of Orthodox ethnicity in the 2013 Bosnian census}
  \item{Macedonian}{The number of people of Macedonian ethnicity in the 2013 Bosnian census}
  \item{Others}{The number of people of Other ethnicity in the 2013 Bosnian census}
  \item{Undeclared}{The number of people who did not declare ethnicity in the 2013 Bosnian census}
  \item{Unknown}{The number of people of Unknown ethnicity in the 2013 Bosnian census}
}

\itemize{
  \item{ons_id}{The Office of National Statistics id for the constituency}
  \item{constituency_name}{The plain language name of the constituency seat}
  \item{first_party}{The parliamentary party with the most votes for a given seat}
  \item{Con}{The number of votes won by the Conservative party for a given seat}
  \item{Lab}{The number of votes won by the Labour party for a given seat}
  \item{LD}{The number of votes won by the Liberal Democrats party for a given seat}
  \item{UKIP}{The number of votes won by the UKIP party for a given seat}
  \item{Green}{The number of votes won by the Green party for a given seat}
}
}
\keyword{datasets}
