% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosnia_shapefile.R, R/london_shapefile.R
\docType{data}
\name{london_shapefile}
\alias{london_shapefile}
\alias{bosnia_shapefile}
\title{Shape Data to replicate Bosnia dot density map from https://twitter.com/ImerM1/status/1037358973807210498}
\format{
A data frame with 141 rows and 2 variables

A data frame with 73 rows and 2 variables
}
\usage{
data(london_shapefile)

data(london_shapefile)
}
\description{
A dataset containing the simple features shapefile for Bosnia & Herzegovina
    districts for the 2013 census. Downloaded from gadm.org

A dataset containing the simple features shapefile for the London
    electoral seats in the 2017 UK General Election
}
\details{
\itemize{
  \item{district}{The name of the district for the 2013 Bosnian census}
  \item{geometery}{The geometry of the shapefile}
}

\itemize{
  \item{ons_id}{The Office of National Statistics id for the constituency}
  \item{geometery}{The geometry of the shapefile}
}
}
\keyword{datasets}
