% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dotdensity.R
\name{theme_dotdensity}
\alias{theme_dotdensity}
\title{theme_dotdensity}
\usage{
theme_dotdensity(legend = TRUE, legend_size = 10,
  background_colour = "#212121", text_colour = "white")
}
\arguments{
\item{legend}{whether or not to include a legend}

\item{legend_size}{the size of the text and icons in the legend}

\item{background_colour}{the colour code of the background of the plot. Defaults to #212121 (very dark grey)}

\item{text_colour}{the text colour for the plot. Defaults to white}
}
\description{
A function which colours the background of a dot density map.
These tend to look best when on a dark background so defaults to darkgrey.
}
\author{
Robert Hickman
}
